Получение арбитражных сделок между биржами

+++ src/main/java/net/protsenko/fundy/app/dto/TickerData.java
@@ -1,13 +1,15 @@
 package net.protsenko.fundy.app.dto;
 
+import java.math.BigDecimal;
+
 public record TickerData(
         TradingInstrument instrument,
-        double lastPrice,
-        double bid,
-        double ask,
-        double high24h,
-        double low24h,
-        double volume24h,
+        BigDecimal lastPrice,
+        BigDecimal bid,
+        BigDecimal ask,
+        BigDecimal high24h,
+        BigDecimal low24h,
+        BigDecimal volume24h,
         long timestamp
 ) {
 }
